merge: fixing caching.ts
fixed all issues with the previous merge commit, also tidied up where certain methods lived
